requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-environ==0.8.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
